Revert "useFetchUsers custom hook"
This reverts commit 83e7264c03d531e9b8a40f2dbea3335fd98cfaf4.
